backend/requirements.dev.txt
anyio
Django~=4.2.13
psycopg
psycopg2-binary
celery~=5.4.0
django-celery-beat
django-celery-results
django-cors-headers
django-filter
django-rest-passwordreset
django-polymorphic
redis
django-redis
djangorestframework
drf-standardized-errors
numpy
Pillow
PyJWT
# django-notifications-hq
django-debug-toolbar
django-environ
django-soft-delete
django-restful-admin
requests
django-debug-toolbar
django-gmailapi-backend
# django-modeltranslation
django-parler
django-rest-framework

stripe==2.55.0
django-crispy-forms
crispy-bootstrap5
django-jet-reboot==1.3.5
django-oscar[sorl-thumbnail]
raven==5.32.0
# supervisor
easy-thumbnails
sorl-thumbnail
whitenoise
django-haystack
Whoosh
pycountry